+++ com.unity.robotics.ros-tcp-connector/Tests/Runtime/RosConnectionTests.cs
@@ -31,5 +31,48 @@ namespace UnitTests
             ROSConnection ros2 = ROSConnection.GetOrCreateInstance();
             Assert.AreEqual(ros, ros2);
         }
+
+        [Test]
+        public void Subscribe_WithLatch_StoresTransientLocalIntentOnTopic()
+        {
+            ROSConnection ros = ROSConnection.GetOrCreateInstance();
+            ros.ConnectOnStart = false;
+            const string topic = "/unit_test_latched_subscription";
+
+            ros.Subscribe<TimeMsg>(topic, _ => { }, latch: true);
+
+            Assert.IsTrue(ros.GetTopic(topic).IsSubscriberLatched);
+        }
+
+        [Test]
+        public void Subscribe_LatchedCallback_UpgradesSharedTopicAndCannotBeDowngraded()
+        {
+            ROSConnection ros = ROSConnection.GetOrCreateInstance();
+            ros.ConnectOnStart = false;
+            const string topic = "/unit_test_latched_subscription_upgrade";
+
+            ros.Subscribe<TimeMsg>(topic, _ => { });
+            Assert.IsFalse(ros.GetTopic(topic).IsSubscriberLatched);
+
+            ros.Subscribe<TimeMsg>(topic, _ => { }, latch: true);
+            Assert.IsTrue(ros.GetTopic(topic).IsSubscriberLatched);
+
+            ros.Subscribe<TimeMsg>(topic, _ => { }, latch: false);
+            Assert.IsTrue(ros.GetTopic(topic).IsSubscriberLatched);
+        }
+
+        [Test]
+        public void Unsubscribe_ClearsLatchedIntentForFutureSubscription()
+        {
+            ROSConnection ros = ROSConnection.GetOrCreateInstance();
+            ros.ConnectOnStart = false;
+            const string topic = "/unit_test_latched_subscription_reset";
+
+            ros.Subscribe<TimeMsg>(topic, _ => { }, latch: true);
+            ros.Unsubscribe(topic);
+            ros.Subscribe<TimeMsg>(topic, _ => { });
+
+            Assert.IsFalse(ros.GetTopic(topic).IsSubscriberLatched);
+        }
     }
 }